Languages

+++ app/src/main/java/com/example/stores/MainActivity.kt
@@ -1,15 +1,21 @@
 package com.example.stores
 
+import android.content.DialogInterface
+import android.content.Intent
+import android.net.Uri
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.widget.Toast
 import androidx.annotation.UiThread
 import androidx.recyclerview.widget.GridLayoutManager
 import com.example.stores.databinding.ActivityMainBinding
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
+import com.google.android.material.snackbar.Snackbar
 import org.jetbrains.anko.doAsync
 import org.jetbrains.anko.runOnUiThread
 import org.jetbrains.anko.uiThread
 
-class MainActivity : AppCompatActivity(), OnClickListener {
+class MainActivity : AppCompatActivity(), OnClickListener , MainAux {
 
     private lateinit var mBinding: ActivityMainBinding
     private lateinit var mAdapter: StoreAdapter
@@ -20,7 +26,7 @@ class MainActivity : AppCompatActivity(), OnClickListener {
         mBinding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(mBinding.root)
 
-        mBinding.btnSave.setOnClickListener {
+        /*mBinding.btnSave.setOnClickListener {
             val store = StoreEntity(name = mBinding.etName.text.toString().trim())
 
             Thread{
@@ -28,9 +34,34 @@ class MainActivity : AppCompatActivity(), OnClickListener {
             }.start()
 
             mAdapter.add(store)
-        }
+        }*/
+
+        mBinding.fab.setOnClickListener { launchEditFragment() }
 
         setupRecyclerView()
+    }
+
+    private fun launchEditFragment(args : Bundle? = null) {
+        val fragment = EditStoreFragment()
+        if (args != null) fragment.arguments = args
+
+        val fragmentManager = supportFragmentManager
+        val fragmentTransaction = fragmentManager.beginTransaction()
+
+        fragmentTransaction.add(R.id.containerMain, fragment)
+        fragmentTransaction.addToBackStack(null)
+        fragmentTransaction.commit()
+
+        //mBinding.fab.hide()
+        hideFab()
+    }
+
+    override fun addStore(storeEntity: StoreEntity) {
+        mAdapter.add(storeEntity)
+    }
+
+    override fun updateStore(storeEntity: StoreEntity) {
+        mAdapter.update(storeEntity)
     }
 
     private fun setupRecyclerView() {
@@ -56,8 +87,11 @@ class MainActivity : AppCompatActivity(), OnClickListener {
     }
 
     // OnClickListener
-    override fun onClick(storeEntity: StoreEntity) {
-        TODO("Not yet implemented")
+    override fun onClick(storeId: Long) {
+        val args = Bundle()
+        args.putLong(getString(R.string.arg_id), storeId)
+
+        launchEditFragment(args)
     }
 
     override fun onFavoriteStore(storeEntity: StoreEntity) {
@@ -66,17 +100,76 @@ class MainActivity : AppCompatActivity(), OnClickListener {
         doAsync{
             StoreApplication.database.storeDao().updateStore(storeEntity)
             uiThread{
-                mAdapter.update(storeEntity)
+                updateStore(storeEntity)
             }
         }
     }
 
     override fun onDeleteStore(storeEntity: StoreEntity) {
-        doAsync {
-            StoreApplication.database.storeDao().deleteStore(storeEntity)
-            uiThread {
-                mAdapter.delete(storeEntity)
+        val items = resources.getStringArray(R.array.array_options_item)
+
+        MaterialAlertDialogBuilder(this)
+            .setTitle(R.string.dialog_options_title)
+            .setItems(items, { dialogInterface, i ->
+                when(i){
+                    0 -> confirmDelete(storeEntity)
+                    1 -> dial(storeEntity.phone)
+                    2 -> goToWebsite(storeEntity.website)
+                }
+            })
+            .show()
+    }
+
+    private fun confirmDelete(storeEntity: StoreEntity){
+        MaterialAlertDialogBuilder(this)
+            .setTitle(R.string.dialog_delete_title)
+            .setPositiveButton(R.string.dialog_delete_confirm, { dialogInterface, i ->
+                doAsync {
+                    StoreApplication.database.storeDao().deleteStore(storeEntity)
+                    uiThread {
+                        mAdapter.delete(storeEntity)
+                    }
+                }
+            })
+            .setNegativeButton(R.string.dialog_delete_cancel, null)
+            .show()
+    }
+
+    private fun dial(phone: String){
+        val callIntent = Intent().apply {
+            action = Intent.ACTION_DIAL
+            data = Uri.parse("Tel:$phone")
+        }
+        startIntent(callIntent)
+    }
+
+    private fun goToWebsite(website: String){
+        if (website.isEmpty()){
+            Toast.makeText(this, R.string.main_error_no_website, Toast.LENGTH_LONG).show()
+        } else {
+            val websiteIntent = Intent().apply {
+                action = Intent.ACTION_VIEW
+                data = Uri.parse(website)
             }
+            startIntent(websiteIntent)
+        }
+    }
+
+    private fun startIntent(intent: Intent){
+        if (intent.resolveActivity(packageManager) != null){
+            startActivity(intent)
+        } else {
+            Toast.makeText(this, R.string.main_error_no_website, Toast.LENGTH_LONG).show()
+        }
+    }
+
+    /*MainAux*/
+
+    override fun hideFab(isVisible: Boolean) {
+        if (isVisible){
+            mBinding.fab.show()
+        } else {
+            mBinding.fab.hide()
         }
     }
 }
